+++ src/client/components/workspace/HistoryOrWorkspaceContainer.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useSelector } from 'react-redux';
 // Import local components
 import BarGraph from '../legacy-components/BarGraph';
-import ScheduleContainer from '../main/loadTest/TestContainer';
+import ScheduleContainer from './TestContainer';
 import WorkspaceContainer from './WorkspaceContainer';
 import HistoryContainer from './HistoryContainer';
 // Import MUI components and icons
